backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
openai==2.7.2
anthropic==0.34.2
pydantic==2.9.2
python-multipart==0.0.6
instaloader==4.10.3
requests==2.31.0
beautifulsoup4==4.14.2
aiohttp>=3.13.0
sqlalchemy>=2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
google-auth==2.27.0
google-genai>=0.2.0
PyPDF2==3.0.1
python-docx==1.1.0
watchdog==3.0.0
hyperspell>=0.1.0
boto3>=1.34.0
mem0ai>=1.0.0
